Arranged code in different file
